% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_inhib_slope}
\alias{calc_inhib_slope}
\title{Calculate the slope of 1/nM.per.hr ~ conc.pNA; return it with pertinent info (e.g p value)}
\usage{
calc_inhib_slope(x)
}
\arguments{
\item{x}{a data frame with nM.per.hr and conc.pNA}
}
\description{
This is just one of three functions that (essentially) replicates the functionality of lm_stats
}
\details{
Caculates the slope (and standard error, rsquared, and pvalue) of inverse activity vs conc.pNA - "m" in equation 4.
}

