% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_pNA_k}
\alias{calc_pNA_k}
\title{Calculates the exponential correction factor for fluorescence quenching by pNA}
\usage{
calc_pNA_k(printInfPlot = FALSE)
}
\arguments{
\item{printInfPlot}{prints plot.}
}
\value{
the exponenetial correction factor $k$
}
\description{
Relies on the data frame pNA_calib.Rdata
}

