% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corr_pNA_conc}
\alias{corr_pNA_conc}
\title{Corrects pNA concentrations for my sloppy initial calculation}
\usage{
corr_pNA_conc(x)
}
\arguments{
\item{x}{data frame with incorrectly calculated pNA concentrations in the `conc.pNA` column}
}
\value{
The data frame with teh conc.pNA column corrected
}
\description{
Initially I foolishly calculated pNA concentrations as if the total volume were only 1 mL. In reality, the total volume was 1 + 0.1 + 0.016 + 0.032
}

