% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corr_stats}
\alias{corr_stats}
\title{Returns useful parameters of a linear model}
\usage{
corr_stats(d, xvar, yvar)
}
\arguments{
\item{d}{Data frame containing the data in question}

\item{xvar}{Name of the x variable (as character)}

\item{yvar}{Name of the y variable (as character)}
}
\value{
A 1-row data frame (no, it makes no sense to do it this way.)
}
\description{
Creates a linear model of the form $y=mx+b$ and returns useful parameters, including:
* slope
* standard error of the slope
* intercept
* standard error of the intercept
* p value
* number of points
  Tries hard to catch errors and return NA in those cases. This is an early version of `lm_stats` in the **enzalyze** package: github.com/adsteen/enzalyze
}

