% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correct_for_pNA_quench}
\alias{correct_for_pNA_quench}
\title{Corrects fluorescence for pNA quenching}
\usage{
correct_for_pNA_quench(x, k)
}
\arguments{
\item{x}{a data frame containing a column called origFl}

\item{k}{an exponential coefficient (e^(kx))}
}
\value{
the same data frame with a corrected column called fl
}
\description{
Based on the model $fl_{corrected} = fl_{obs} \times e^{k\times[\text{pNA}]}$
}

