% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_homologous}
\alias{find_homologous}
\title{Finds pNA substrates homologous to a AMC substrates (i.e., Leu-pNA for Leu-AMC)}
\usage{
find_homologous(x)
}
\arguments{
\item{x}{A data frame with columns pNA.subs and AMC.substrate (you like the antiparallelism?)}
}
\value{
The same data frame with a new column, `is.homologous`, with boolean for whether `pNA.subs` is homologous to `AMC.substrate`
}
\description{
This function is terrible and if you are paying attention to it you should re-evaluate the choices you have made in life.
}

