% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gauss}
\alias{gauss}
\title{calculate a Gaussian distribution}
\usage{
gauss(x, mu, sigma)
}
\arguments{
\item{x}{vector of values on which to calculate the distribution}

\item{mu}{Mean}

\item{sigma}{standard deviation}
}
\description{
calculate a Gaussian distribution
}
\details{
You would think there would be a function for this in base R, but if there is I can't find it
}

