% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_nls_coefs}
\alias{get_nls_coefs}
\title{Gets coefficients and standard errors for an nls object}
\usage{
get_nls_coefs(x)
}
\arguments{
\item{x}{an nls object}
}
\value{
a 1-row data frame. Yes, this is dumb: it should be a vector
}
\description{
Gets coefficients and standard errors for an nls object
}

