% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm_stats}
\alias{lm_stats}
\title{Returns statistics for linear regressions (e.g. slope, intercept, error, etc)}
\usage{
lm_stats(d, xvar, yvar, debug = FALSE)
}
\arguments{
\item{d}{Data frame containing at least two numeric columns}

\item{xvar}{Numeric column to be used as the independant variable for the regression}

\item{yvar}{Numeric column to be used as the dependant variable for the regression}
}
\value{
Returns a one-row data frame containing (at present): slope, intercept, slope standard error, intercept standard error, p value, r-squared, and number of points.
}
\description{
Returns statistics for linear regressions (e.g. slope, intercept, error, etc)
}
\details{
Creates linear models of a single independant variable
}

