% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_sat_curves}
\alias{make_sat_curves}
\title{Analyzes and plots all saturation curves}
\usage{
make_sat_curves(print_plot = TRUE, save_plot = FALSE, fn = NA)
}
\arguments{
\item{print_plot}{Whether to print plots}

\item{save_plot}{Whether to save plots}

\item{fn}{filename for saved plots}
}
\description{
Relies on 4 data files in /data/: sat_arg_TN, sat_leu_TN, sat_curves_IMS, sat_curve_calib
}

