% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p_val_labeller}
\alias{p_val_labeller}
\title{Returns nicely formatted p values}
\usage{
p_val_labeller(pval, sigfigs = 2, cutoff = 0.001, plotmath = TRUE)
}
\arguments{
\item{pval}{A vector of pvalues}

\item{sigfigs}{The number of significant figures to which to round each p value}

\item{cutoff}{The number below which the pvalue will be reported as "p<cutoff"}

\item{plotmath}{Whether to format for plotmath or plain text. Plotmath uses "p==", which will display onscreen as "p=" if it is interpreted as plotmath}
}
\description{
Returns nicely formatted p values
}

