% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_affinity_corr}
\alias{plot_affinity_corr}
\title{Make plots of affinity correlations to DI, MW, and delta G of reaction}
\usage{
plot_affinity_corr(x, label_size = 2, hjust = 0, vjust = 0, sz = 0.25,
  inhib = 4, letter_size = 3, ...)
}
\arguments{
\item{x}{should be `inhib_data` from script}

\item{label_size}{}

\item{print_plots}{NOT USED}

\item{save_plots}{NOT USED}

\item{fn}{for saved plots NOT USED}
}
\description{
Make plots of affinity correlations to DI, MW, and delta G of reaction
}

