% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_all_inv_v0}
\alias{plot_all_inv_v0}
\title{Makes fig 4}
\usage{
plot_all_inv_v0(d, print_plot = TRUE, save_plot = FALSE, fn = NA,
  height = 4, width = 7, dpi = 300, ...)
}
\arguments{
\item{d}{data frame}
}
\description{
Makes fig 4
}

