% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_KI_norm_vs_Xaa_pNA}
\alias{plot_KI_norm_vs_Xaa_pNA}
\title{Make a faceted dotplot of KI/KI,ref for each Xaa-pNA, Yaa-AMC and location}
\usage{
plot_KI_norm_vs_Xaa_pNA(x, vertical = FALSE, print_plot = TRUE,
  save_plot = TRUE, fn = NA, sz = 0.25, ...)
}
\arguments{
\item{x}{The underlying data (all_rel_Ki)}

\item{vertical}{Whether to build a vertical or gridded plot (vertical is not up-to-date)}

\item{print_plot}{Whether to print the plot}

\item{save_plot}{Whether to save the plot}

\item{fn}{Filename for the saved plot}

\item{sz}{'Size' (i.e. thickness) of the lines}
}
\value{
The plot object, which is discarded by write_paper()
}
\description{
Make a faceted dotplot of KI/KI,ref for each Xaa-pNA, Yaa-AMC and location
}

