% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_inhib_data}
\alias{read_inhib_data}
\title{Reads in data frames, corrects bad data, and returns a master list of the data}
\usage{
read_inhib_data(printDataCheck = FALSE)
}
\arguments{
\item{printDataCheck}{prints plots of the data, to show what bad data points were thrown out.}
}
\value{
a list containing each data set
}
\description{
Reads in data frames, corrects bad data, and returns a master list of the data
}

