% Generated by roxygen2 (4.0.2): do not edit by hand
\name{slope_and_SE}
\alias{slope_and_SE}
\title{Calculate the slope and standard error of fluorescence vs incubation time}
\usage{
slope_and_SE(x)
}
\arguments{
\item{x}{is a data frame containing columns `fl` and `incTime` (both numerics)}
}
\description{
Calculate the slope and standard error of fluorescence vs incubation time
}

