% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_paper}
\alias{write_paper}
\title{Master function to replicate all analyses in the manuscript}
\usage{
write_paper(path = "", print_plots = TRUE, save_plots = FALSE)
}
\arguments{
\item{path}{Where to save plot files}

\item{print_plots}{Whether to print plots to the default graphics device (usually your screen)
Note that this may have no effect in command-line R.}

\item{save_plots}{Whether to save plots.}
}
\value{
A data frame with everthing calculated in Fig 7
}
\description{
Reproduces data analysis from Steen et al (2015) AME paper about peptidase substrate specificity
}
\details{
Running this as write_paper(print_plots=TRUE, save_plots=FALSE) *should* work and be platform-independant. Beyond that I don't really make any guarantees but I'm happy to help via email. Notably, save_plots=TRUE requires the existence of a subdirectory of the working directory named `plots` and may be platform-dependant (this was built using OSX 10.10.4 and the cairo device). If you wish to take apart my analysis to understand it better or show that it is totally bogus, I recommend copy-pasting the inside of this function from R and running it as a script.
}

